+++ flutter_inappwebview_android/lib/src/in_app_webview/in_app_webview_controller.dart
@@ -2096,6 +2096,11 @@ class AndroidInAppWebViewController extends PlatformInAppWebViewController
   }
 
   @override
+  Future<void> waitForInitialJavaScriptBridgeReady() async {
+    await channel?.invokeMethod<void>('waitForInitialJavaScriptBridgeReady');
+  }
+
+  @override
   Future<void> postUrl({
     required WebUri url,
     required Uint8List postData,